feat: ability to see the size difference between two package versions
