+++ report.go
@@ -5,6 +5,7 @@ import (
 	"math/big"
 	"package_size_calculator/pkg/npm"
 	"package_size_calculator/pkg/time_helpers"
+	"strings"
 	"time"
 
 	"github.com/dustin/go-humanize"
@@ -28,46 +29,30 @@ func printReport(
 	addedDependencies []*npm.PackageJSON,
 	deps map[string]*dependencyPackageInfo,
 ) {
-	packageInfo := modifiedPackage.Info
 	package_ := modifiedPackage.Package
 	packageJson := package_.JSON
-	pkgDownloads := modifiedPackage.DownloadsLastWeek
-	installedPackageSize := modifiedPackage.Size
+	downloadsLastWeek := modifiedPackage.DownloadsLastWeek
+	oldPackageSize := modifiedPackage.Size
 
 	modifiedPackageName := boldYellow.Sprint(packageJson.String())
 
-	estPackageSize := installedPackageSize
-	packageSizeWithoutRemovedDeps := installedPackageSize
+	newPackageSize := oldPackageSize
+	packageSizeWithoutRemovedDeps := oldPackageSize
 	for _, p := range deps {
 		if p.Type == DependencyRemoved {
-			estPackageSize -= p.Size
+			newPackageSize -= p.Size
 			packageSizeWithoutRemovedDeps -= p.Size
 		} else {
-			estPackageSize += p.Size
+			newPackageSize += p.Size
 		}
 	}
 
 	fmt.Println()
 	boldGreen.Println("Package size report")
 	boldGreen.Println("===================")
-	fmt.Println()
-	fmt.Printf("%s: %s\n", bold.Sprintf("Package info for \"%s\"", modifiedPackageName), humanize.Bytes(installedPackageSize))
-	fmt.Printf(
-		"  %s: %s %s\n",
-		bold.Sprint("Released"),
-		package_.ReleaseTime,
-		grayParens("%s ago", time_helpers.FormatDuration(time.Since(package_.ReleaseTime))),
-	)
-	fmt.Printf("  %s: %s\n", bold.Sprint("Downloads last week"), fmtInt(int(pkgDownloads)))
-	fmt.Printf("  %s: %s\n", bold.Sprint("Estimated traffic last week"), humanize.Bytes(pkgDownloads*installedPackageSize))
 
-	if packageJson.Version != packageInfo.LatestVersion.JSON.Version {
-		fmt.Printf("  %s: %s %s\n",
-			bold.Sprint("Latest version"),
-			packageInfo.LatestVersion.Version,
-			grayParens("%s ago", time_helpers.FormatDuration(time.Since(packageInfo.LatestVersion.ReleaseTime))),
-		)
-	}
+	fmt.Println()
+	reportPackageInfo(modifiedPackage, true, 0)
 
 	if len(removedDependencies) > 0 {
 		fmt.Println()
@@ -76,22 +61,22 @@ func printReport(
 		for _, p := range removedDependencies {
 			info := deps[p.String()]
 
-			pcSize := float64(info.Size) * 100 / float64(installedPackageSize)
+			pcSize := float64(info.Size) * 100 / float64(oldPackageSize)
 			traffic := info.DownloadsLastWeek * info.Size
-			pcTraffic := float64(pkgDownloads) * 100 / float64(info.DownloadsLastWeek)
+			pcTraffic := float64(downloadsLastWeek) * 100 / float64(info.DownloadsLastWeek)
 
 			fmt.Printf("  %s %s: %s %s\n", color.RedString("-"), boldYellow.Sprint(p.String()), humanize.Bytes(info.Size), grayParens("%s%%", fmtPercent(pcSize)))
 			fmt.Printf("    %s: %s\n", bold.Sprint("Downloads last week"), fmtInt(int(info.DownloadsLastWeek)))
 			fmt.Printf(
 				"    %s: %s %s\n",
 				bold.Sprintf("Downloads last week from \"%s\"", modifiedPackageName),
-				fmtInt(int(pkgDownloads)),
+				fmtInt(int(downloadsLastWeek)),
 				grayParens("%s%%", fmtPercent(pcTraffic)),
 			)
 			fmt.Printf("    %s: %s\n", bold.Sprint("Estimated traffic last week"), humanize.Bytes(traffic))
 			fmt.Printf("    %s: %s %s\n",
 				bold.Sprintf("Estimated traffic from \"%s\"", modifiedPackageName),
-				humanize.Bytes(pkgDownloads*info.Size),
+				humanize.Bytes(downloadsLastWeek*info.Size),
 				grayParens("%s%%", fmtPercent(pcTraffic)),
 			)
 		}
@@ -118,19 +103,59 @@ func printReport(
 		}
 	}
 
+	fmt.Println()
+	reportSizeDifference(oldPackageSize, newPackageSize, downloadsLastWeek)
+}
+
+func reportPackageInfo(modifiedPackage *packageInfo, showLatestVersionHint bool, indentation int) {
+	indent := strings.Repeat(" ", indentation)
+
+	packageInfo := modifiedPackage.Info
+	package_ := modifiedPackage.Package
+	packageJson := package_.JSON
+	downloadsLastWeek := modifiedPackage.DownloadsLastWeek
+	oldPackageSize := modifiedPackage.Size
+
+	modifiedPackageName := boldYellow.Sprint(packageJson.String())
+
+	fmt.Printf("%s%s: %s\n", indent, bold.Sprintf("Package info for \"%s\"", modifiedPackageName), humanize.Bytes(oldPackageSize))
+	fmt.Printf(
+		"%s  %s: %s %s\n",
+		indent,
+		bold.Sprint("Released"),
+		package_.ReleaseTime,
+		grayParens("%s ago", time_helpers.FormatDuration(time.Since(package_.ReleaseTime))),
+	)
+	fmt.Printf("%s  %s: %s\n", indent, bold.Sprint("Downloads last week"), fmtInt(int(downloadsLastWeek)))
+	fmt.Printf("%s  %s: %s\n", indent, bold.Sprint("Estimated traffic last week"), humanize.Bytes(downloadsLastWeek*oldPackageSize))
+
+	if showLatestVersionHint {
+		latestVersion := packageInfo.LatestVersion
+		if packageJson.Version != latestVersion.JSON.Version {
+			fmt.Printf("%s  %s: %s %s\n",
+				indent,
+				bold.Sprint("Latest version"),
+				latestVersion.Version,
+				grayParens("%s ago", time_helpers.FormatDuration(time.Since(latestVersion.ReleaseTime))),
+			)
+		}
+	}
+}
+
+func reportSizeDifference(oldSize, newSize, downloads uint64) {
 	indicatorColor := boldGreen
-	if estPackageSize > installedPackageSize {
+	if newSize > oldSize {
 		indicatorColor = boldRed
-	} else if estPackageSize == installedPackageSize {
+	} else if newSize == oldSize {
 		indicatorColor = boldGray
 	}
 
-	pcSize := 100 * float64(estPackageSize) / float64(installedPackageSize)
+	pcSize := 100 * float64(newSize) / float64(oldSize)
 	pcSizeFmt := indicatorColor.Sprintf("%s%%", fmtPercent(pcSize))
 
-	oldTrafficLastWeek := big.NewInt(int64(pkgDownloads * installedPackageSize))
+	oldTrafficLastWeek := big.NewInt(int64(downloads * oldSize))
 	oldTrafficLastWeekFmt := humanize.BigBytes(oldTrafficLastWeek)
-	estTrafficNextWeek := big.NewInt(int64(pkgDownloads * estPackageSize))
+	estTrafficNextWeek := big.NewInt(int64(downloads * newSize))
 	estTrafficNextWeekFmt := humanize.BigBytes(estTrafficNextWeek)
 
 	estTrafficChange := big.NewInt(0).Sub(oldTrafficLastWeek, estTrafficNextWeek)
@@ -145,13 +170,12 @@ func printReport(
 	}
 	estTrafficChangeFmt = indicatorColor.Sprintf(estTrafficChangeFmt, humanize.BigBytes(estTrafficChange))
 
-	fmt.Println()
 	fmt.Printf(
 		"%s: %s %s %s %s\n",
 		bold.Sprint("Estimated package size"),
-		humanize.Bytes(installedPackageSize),
+		humanize.Bytes(oldSize),
 		arrow,
-		indicatorColor.Sprintf(humanize.Bytes(estPackageSize)),
+		indicatorColor.Sprintf(humanize.Bytes(newSize)),
 		grayParens("%s", pcSizeFmt),
 	)
 	fmt.Printf(
@@ -162,7 +186,6 @@ func printReport(
 		indicatorColor.Sprint(estTrafficNextWeekFmt),
 		grayParens("%s", estTrafficChangeFmt),
 	)
-	fmt.Println()
 }
 
 func grayParens(s string, args ...any) string {
@@ -170,4 +193,12 @@ func grayParens(s string, args ...any) string {
 	b := gray.Sprint(")")
 
 	return fmt.Sprintf("%s%s%s", a, fmt.Sprintf(s, args...), b)
+}
+
+func fmtPercent(v float64) string {
+	return humanize.FormatFloat("#,###.##", v)
+}
+
+func fmtInt(v int) string {
+	return humanize.FormatInteger("#,###.", v)
 }
